+++ CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/).
 
 ## Unreleased
+### Added
+- `e d` to edit entry description
+- `e t` to edit entry task
+- `e p` to edit entry project
+
+### Changed
+- `reaper-edit-entry` moved from `e` to `e e`
 
 ## 1.2.1 - 2020-11-14
 ### Fixed